refactor of pipelineManager code for threadsafety
